virtiofsd: Build inside a container

+++ tools/packaging/static-build/virtiofsd/build-static-virtiofsd.sh
@@ -16,10 +16,13 @@ script_dir="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
 source "${script_dir}/../../scripts/lib.sh"
 
+virtiofsd_repo="${virtiofsd_repo:-}"
 virtiofsd_version="${virtiofsd_version:-}"
+virtiofsd_zip="${virtiofsd_zip:-}"
 
-[ -n "$virtiofsd_version" ] || virtiofsd_version=$(get_from_kata_deps "externals.virtiofsd.version")
+[ -n "$virtiofsd_repo" ] || die "failed to get virtiofsd repo"
 [ -n "$virtiofsd_version" ] || die "failed to get virtiofsd version"
+[ -n "${virtiofsd_zip}" ] || die "failed to get virtiofsd binary URL"
 
 [ -d "virtiofsd" ] && rm -r virtiofsd
 
@@ -28,8 +31,6 @@ pull_virtiofsd_released_binary() {
 	info "Only x86_64 binaries are distributed as part of the virtiofsd releases" && return 1
     fi
     info "Download virtiofsd version: ${virtiofsd_version}"
-    virtiofsd_zip=$(get_from_kata_deps "externals.virtiofsd.meta.binary")
-    [ -n "${virtiofsd_zip}" ] || die "failed to get virtiofsd binary URL"
 
     mkdir -p virtiofsd
 
@@ -44,31 +45,35 @@ pull_virtiofsd_released_binary() {
 }
 
 init_env() {
+   source "$HOME/.cargo/env"
+
    case ${ARCH} in
      "aarch64")
        LIBC="musl"
+       ARCH_LIBC=""
      ;;
      "ppc64le")
        LIBC="gnu"
        ARCH="powerpc64le"
+       ARCH_LIBC=${ARCH}-linux-${LIBC}
      ;;
      "s390x")
        LIBC="gnu"
+       ARCH_LIBC=${ARCH}-linux-${LIBC}
      ;;
      "x86_64")
        LIBC="musl"
+       ARCH_LIBC=""
+     ;;
   esac
 
-    ARCH_LIBC=${ARCH}-linux-${LIBC}
 }
 	   
 build_virtiofsd_from_source() {
    echo "build viriofsd from source"
    init_env
 
-   virtiofsd_url=$(get_from_kata_deps "externals.virtiofsd.url")
-
-   git clone --depth 1 --branch ${virtiofsd_version} ${virtiofsd_url} virtiofsd
+   git clone --depth 1 --branch ${virtiofsd_version} ${virtiofsd_repo} virtiofsd
    pushd virtiofsd
 
    export RUSTFLAGS='-C target-feature=+crt-static -C link-self-contained=yes'
